Update structure
And fix auto center problem for inforwindow
