+++ src/components/InfoWindow.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState }  from "react";
+import React from "react";
 import { Card } from "antd";
 import {CustomOverlay} from 'react-bmapgl';
