+++ src/components/index.js
@@ -2,3 +2,4 @@
 export { default as InfoHeader } from "./InfoHeader";
 export { default as InfoMarker } from "./InfoMarker";
 export { default as InfoWindow } from "./InfoWindow";
+export { default as BaiduMap } from "./BaiduMap";
